requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-jose
alembic
pydantic
pydantic-settings
python-multipart
httpx
asyncpg
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
redis==5.0.3
aiofiles>=23.2.1
python-multipart>=0.0.9
pytest-asyncio
pytest-cov
aiosqlite>=0.20
